Build submit workflow entry page + routing
